+++ Examples/Rot3DLib/CmdParameter/Support/Strings.cpp
