+++ Examples/Rot3DLib/Parameters.cpp
@@ -19,7 +19,6 @@
 #include <cassert>
 #include <sstream>
 #include <iostream>
-#include "Support/CmdParameter.h"
 
 using namespace std;
