+++ Makefile
@@ -105,12 +105,12 @@ EXAMPLE_TARGETS = $(patsubst %,$(OBJ_DIR)/bin/%,$(EXAMPLES))
 ROT3DLIB_OBJ := \
 	Rot3DLib/Rot3DKernels.o    \
 	Rot3DLib/Parameters.o      \
-	Rot3DLib/Support/Support/Strings.o \
-	Rot3DLib/Support/Types/IntParameter.o \
-	Rot3DLib/Support/Types/NoneParameter.o \
-	Rot3DLib/Support/Types/Types.o \
-	Rot3DLib/Support/DefParameter.o \
-	Rot3DLib/Support/CmdParameter.o
+	Rot3DLib/CmdParameter/Support/Strings.o \
+	Rot3DLib/CmdParameter/Types/IntParameter.o \
+	Rot3DLib/CmdParameter/Types/NoneParameter.o \
+	Rot3DLib/CmdParameter/Types/Types.o \
+	Rot3DLib/CmdParameter/DefParameter.o \
+	Rot3DLib/CmdParameter/CmdParameter.o
 
 
 ROT3DLIB_OBJ := $(patsubst %,$(OBJ_DIR)/Examples/%,$(ROT3DLIB_OBJ))
@@ -266,7 +266,7 @@ $(OBJ_DIR):
 	@mkdir -p $(OBJ_DIR)/Source
 	@mkdir -p $(OBJ_DIR)/Target
 	@mkdir -p $(OBJ_DIR)/VideoCore
-	@mkdir -p $(OBJ_DIR)/Examples/Rot3DLib/Support/Support  # Creates all directories in path as well
-	@mkdir -p $(OBJ_DIR)/Examples/Rot3DLib/Support/Types
+	@mkdir -p $(OBJ_DIR)/Examples/Rot3DLib/CmdParameter/Support  # Creates all directories in path as well
+	@mkdir -p $(OBJ_DIR)/Examples/Rot3DLib/CmdParameter/Types
 	@mkdir -p $(OBJ_DIR)/Tools
 	@mkdir -p $(OBJ_DIR)/bin
